Deleting the old JS files to upload the new ones.
I have changed the JS implementation considerably. Various JS files are going to contain different globals, and they are going to reference one another. I need them to be in the right order, otherwise global code will get messed up. It is best that the JS files are numbered, given the way the build script works.